compaction: refactor compaction_fwd.hh out of compaction_descriptor.hh
So it can be used in the next patch that will refactor
compaction_state out of class compaction_manager.

(cherry picked from commit 73280c0a15b9160ae175f2a17ed99ed4ec6afc90)
